+++ d2/src/main/java/com/linkedin/d2/balancer/D2ClientConfig.java
@@ -191,6 +191,7 @@ public class D2ClientConfig
   public XdsClientOtelMetricsProvider xdsClientOtelMetricsProvider = new NoOpXdsClientOtelMetricsProvider();
   public boolean loadBalanceStreamException = false;
   public boolean enablePotentialClientsCache = false;
+  public boolean enableRelativeStrategyDeferredAllocation = false;
   public boolean xdsInitialResourceVersionsEnabled = false;
   public Integer xdsStreamMaxRetryBackoffSeconds = null;
   public String xdsMinimumJavaVersion = DEFAULT_MINIMUM_JAVA_VERSION;